Also compute a `unique` count based on the user IP
As discussed on Bundler's slack. The usage statistics are not very
usable, because a small minority of users can generate lots of traffic
and trust the top of the stats.

This for instance happens with large fleets of AWS EC2 instances running
older Linux distributions with the system Ruby 2.3, causing 2.3 to be almost
half of all requests.

To counteract this, we can for each metric only count an IP once. Meaning
some kind of automated script doing 1k requests/day will count for 1, just
like a human user doing 5 requests/day.

The "unicity key" is `[user_agent, client_ip, day]`, so a single IP doing requests
with multiple ruby or bundler versions will be counted once per day in the respective buckets.

In essence it's the same idea than "daily unique visitor" kind of metrics.

This should hopefully give a more relevant picture of Ruby and Bundler
version usage.