avoid contending on single process-wide precompiled regexes
precompiling the `Regex` instances is good! but being behind
`lazy_static!` means all threads are contending an implicit futex
somewhere in the lazy_static implementation. on a 32-core system with 32
logs to process in parallel, this ends up with the top symbol in
`perf stat` to be something in `futex` and an awful lot of total time
waiting on _system stuff_ for no apparent reason.

having set of Regex bundles per-thread keeps contention down at the
cost of a bit more init on the multi-processor case, but overall at
least on my system brings a 31-file benchmark down from 4.6 seconds to
about 2.2 seconds.